Add file closure to fchkparser type
Files open as fchkparser instances can be close
through the close() method.
The files are now closed as soon as data have been
extracted.  This seems to fix an unrelated bug
with GFortran and "-O4" where wrong values were
printed in G if SCR files were constructed (to
compare with SOS/GUVCDE.)
An exception for file operation has been also
added to the "exception" module.