MCD: added flag to add/ignore <ia|O|ja> term
There is a strange discrepancy between closed and open shell formalisms
in GUVCDE, where the prefactor term to compute excited-to-excited
transition moments ignores some of the <ia|O|ja> contributions in the
case of closed-shell but not open-shell.
Comparing unrestricted and restricted 50-50, this discrepancy does not
seem to be correct and make little sense theoretically.  For now, for
further testing, users can decide to include or not this term with
`--debug=ijaa` (default) or ignore with `--debug==noijaa`