Updated atomic database and physical constants
The atomic database has been extensively rewritten to support more
quantities and sources from ESTAMPES have been reviewed and some data
revised.
To avoid having to duplicate the work between ELEMENTS and ESTAMPES, the
database in the former is considered as the reference, and a program,
`gen_py_atom` can generate a Python source code to be included in
ESTAMPES (`data/atom.py`) with updated data.

In order to provide more consistent data in the atomic database, the
physics module has also been revised; physical constants are all
provided as constants at the module level, and the type PhysFact simply
offers a container to conversion functions.