Added segmented loading of molecular data
`build_moldata` loaded systematically all data, which may not be
systematically necessary for some applications, in particular basis set
and molecular orbitals specifications.
The procedure can be called multiple times and will do a brief test that
data seem consistent, at least in terms of number of atoms and
electrons.

A new exception, `QuantityError`, has been added in this occasion.