Software-specific parsing code in new parsers dir.
To improve the clarity of the code and render the core/ content more
general, the parsing tools for files generated by specific tools (e.g.,
Gaussian) are moved to a new directory in `src/` named `parsers/`.
In practice, this would mean that any patch/update regarding the process
of data extraction itself should involved only the files in this
directory, the rest not being impacted.