New get_data method in DataFile to extract data
The new get_data method provides a simple way to extract specific data.
It copies what ESTAMPES does with DataFile.get_data, but the logic here
is simpler since it is assumed the needs in ELEMENTS are more basic.
The code is prototypical and may evolve with time and need.
Only a few properties are supported for now.