Better interface to `Eckart_orient` in `geometry`
The older version of the interface was problematic because the
built-in possibility to override the starting data set required
quantities to have an intent(inout), which was a problem when called by
routines who took arguments with pure `intent(in)` (not requiring an
overwriting of the original data).
To avoid this problem, the original routines are now "pure" in the sense
they do not expect any overlap between initial and final data structures
if the caller asked to store the final structures.  New routines were
created to handle explicitly the overlap case, with a mandatory keyword
to indicate if the starting data set should be overwritten,