New error handler in exception
The new error handler type, `ErrorHandler` provides a simpler and
automated way to raise errors and warnings, and handle the program
execution when encountering this kind of issues.
The module also provides an instance of this type, `runstat`, to be used
by the program units.
The new handler coexists with the existing Error types, which are more
intended for operations on objects, while the new handler provides a
simpler abstraction for standard procedures.

Note: The compilation sequence of corelib in xmake had to be change to
reflect the dependency of `exception` on `output`.