Fixed missing Datafile close after program check
When checking the file type and program version, the data file was
opened twice and only closed once.  This resulted in files being opened
twice at the same time when reading a file, once with an old instance,
and once with a newer instance.
The unnecessary file opening is removed and the file properly closed
after checking the file type and program version.