Adjustable leading blank characters in headers
Headers were originally designed to have a leading blank character.
However, this made it more difficult for some programs to get a proper
layout.  The default now is to not have leading blank characters but
users can set the number of leading characters with the optional
keyword `lead_spaces` in `sec_header`.

Note that to be able to consider also the case of no spacing, tX is used
since the number must be strictly positive.