convert_pure2cart_bsetDB2DB now updates n_basis
`convert_pure2cart_bsetDB2DB` in module `basisset`/`basisset_purecart`
now correctly updates `n_basis`.

Note that currently there is no way to fix `n_basok` if there are
linearly dependent pure functions.  For this reason, the routine
sets it to 0 so it cannot be used with wrong results.  Otherwise,
the same value as `n_basis` is reported.